requirements.txt
catboost==1.0.3
category-encoders==2.3.0
geopy==2.2.0
libzero==0.0.8
lightgbm==3.3.1
numpy==1.21.4
optuna==2.9.1
pandas==1.3.4
pyarrow==6.0.0
rtdl==0.0.9
scikit-learn==1.0.1
scipy==1.8.
tensorboard==2.7.0
tomli==1.2.2
tomli-w==1.0.0
torchinfo==1.5.3
tqdm==4.62.3
xgboost==1.5.0

